+++ app/webpack/webpack.common.js
@@ -2,6 +2,13 @@ const path = require('path');
 const { EnvironmentPlugin } = require('webpack');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+const {
+  ENABLE_TRACKING,
+  GTM_CONTAINER_ID,
+  GA_ID,
+} = process.env;
 
 module.exports = ({ SOURCES_PATH, OUTPUT_PATH, PUBLIC_PATH }) => ({
 
@@ -74,6 +81,15 @@ module.exports = ({ SOURCES_PATH, OUTPUT_PATH, PUBLIC_PATH }) => ({
       patterns: [
         path.join(PUBLIC_PATH, 'robots.txt'),
       ],
+    }),
+
+    new HtmlWebpackPlugin({
+      template: path.join(PUBLIC_PATH, 'index.ejs'),
+      templateParameters: {
+        ENABLE_TRACKING,
+        GTM_CONTAINER_ID,
+        GA_ID,
+      },
     }),
 
   ],