+++ app/CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Place a border around the integration
 
+### Fixed
+
+- Track a single custom event
+
 ## [0.4.2] - 2020-06-14
 
 ### Added